pyproject.toml
[project]
name = "ctic-curriculum-engine"
version = "0.1.0"
description = "CTIC Curriculum Recommendation Engine - Data Pipeline"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12",
    "requests>=2.31",
    "psycopg[binary]>=3.1",
    "pgvector>=0.3",
    "sentence-transformers>=3.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "google-genai>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
